+++ include/whiskey/Parsing/EvalLiterals.hpp
@@ -11,7 +11,7 @@ void appendChar16ToString8(std::string &value, char16_t chr);
 void appendChar32ToString8(std::string &value, char32_t chr);
 void appendChar32ToString16(std::u16string &value, char32_t chr);
 
-bool evalLiteralUInt(const Range &range, UInt &value);
+bool evalLiteralUInt(const Range &range, UInt64 &value);
 bool evalLiteralReal(const Range &range, Real &value);
 bool evalLiteralChar(const Range &range, Char32 &value);
 bool evalLiteralString8(const Range &range, std::string &value);
